requirements.txt
altair==6.1.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.47
idna==3.13
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markupsafe==3.0.3
matplotlib==3.10.9
mpmath==1.3.0
narwhals==2.20.0
networkx==3.6.1
numpy==2.4.3
packaging==26.2
pandas==3.0.2
pillow==12.1.1
protobuf==7.34.1
pyarrow==24.0.0
pydeck==0.9.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.3
streamlit==1.56.0
sympy==1.13.1
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.5
tqdm==4.67.3
typing-extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
watchdog==6.0.0
